pyproject.toml
[project]
name = "lora-hardware-model"
version = "0.1.0"
description = "Python class for lora tunning firmware"
authors = [
    {name = "Yurii Shkoropad",email = "lvivgeorge@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial (>=3.5,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
